Delivery_Agent_Project/requirements.txt
fastapi
uvicorn
streamlit
pandas
requests
torch
torchvision
pillow
numpy